Move routes to theme
